Feature: Add methods to retrieve comments by ID & user ID

+++ src/model/comment.model.js
@@ -26,21 +26,21 @@ const commentSchema = new Schema(
  * @desc Create a new comment
  */
 commentSchema.statics.createComment = async function (userId, postId, text) {
-	return await this.create({ userId, postId, text });
+	return await this.create({ userId: userId, postId: postId, text: text });
 };
 
 /**
  * @desc Delete a comment
  */
 commentSchema.statics.deleteComment = async function (commentId, userId) {
-	return await this.findOneAndDelete({ _id: commentId, userId });
+	return await this.findOneAndDelete({ _id: commentId, userId: userId });
 };
 
 /**
  * @desc Get all comments for a post
  */
 commentSchema.statics.getCommentsByPost = async function (postId) {
-	return await this.find({ postId })
+	return await this.find({ postId: postId })
 		.populate("userId", "username profilePicture")
 		.sort({ createdAt: -1 });
 };
@@ -49,7 +49,14 @@ commentSchema.statics.getCommentsByPost = async function (postId) {
  * @desc Get total comments count for a post
  */
 commentSchema.statics.getCommentsCount = async function (postId) {
-	return await this.countDocuments({ postId });
+	return await this.countDocuments({ postId: postId });
+};
+
+/**
+ * @desc Get Comment By Id and UserId
+ */
+commentSchema.statics.getCommentByIdAndUserId = async function (commentId, userId) {
+    return await this.findOne({ _id: commentId, userId: userId });
 };
 
 module.exports = mongoose.model("Comment", commentSchema);
